# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Agents Management Platform API
  description: |
    Enterprise-grade API for managing AI agents, executions, tools, and analytics.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    ## Rate Limiting
    - API endpoints: 100 requests per minute
    - Auth endpoints: 5 requests per minute
    
    ## WebSocket
    Real-time updates available at `/socket.io` for run streaming and notifications.
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /agents:
    get:
      tags: [Agents]
      summary: List agents
      operationId: listAgents
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AgentStatus'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Agents]
      summary: Create agent
      operationId: createAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentDto'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent by ID
      operationId: getAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Agents]
      summary: Update agent
      operationId: updateAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentDto'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Agents]
      summary: Delete agent
      operationId: deleteAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{id}/clone:
    post:
      tags: [Agents]
      summary: Clone agent
      operationId: cloneAgent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Agent cloned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'

  # ==================== RUNS ====================
...
  /analytics/dashboard:
    get:
      tags: [Analytics]
      summary: Get dashboard data
      operationId: getDashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardData'
  /analytics/tokens:
    get:
      tags: [Analytics]
      summary: Get token usage
      operationId: getTokenUsage
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Token usage data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenUsage'

  # ==================== AUDIT ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    AgentStatus:
      type: string
      enum: [active, inactive, draft]
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        model:
          type: string
        systemPrompt:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        status:
          $ref: '#/components/schemas/AgentStatus'
        version:
          type: integer
        tools:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAgentDto:
      type: object
      required: [name, model]
      properties:
        name:
          type: string
        description:
          type: string
        model:
          type: string
        systemPrompt:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: integer
        status:
          $ref: '#/components/schemas/AgentStatus'
        tools:
          type: array
          items:
            type: string
    UpdateAgentDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        model:
          type: string
        systemPrompt:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        status:
          $ref: '#/components/schemas/AgentStatus'
        tools:
          type: array
          items:
            type: string
    AgentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Runs
...
    DashboardData:
      type: object
      properties:
        overview:
          type: object
          properties:
            totalAgents:
              type: integer
            activeAgents:
              type: integer
            totalRuns:
              type: integer
            runsToday:
              type: integer
    TokenUsage:
      type: object
      properties:
        total:
          type: integer
        byProvider:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              tokens:
                type: integer
              cost:
                type: number

    # Audit
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
security:
  - bearerAuth: []